requirements.txt
firebase-admin==6.5.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.134.0
python-dotenv==1.0.1
email-validator==2.1.1
colorama==0.4.6
pytest==8.2.2
pytest-cov==5.0.0